+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * `ShippingQueryIdentifier` has been renamed to `ShippingQueryId` and now is `value class`
   * `Identifier` has been renamed to `RawChatId` and now is `value class`
   * `ShippingOptionIdentifier` has been renamed to `ShippingOptionId` and now is `value class`
+  * `PollIdentifier` has been renamed to `PollId` and now is `value class`
 
 ## 10.1.2